refactor : 로그인, 회원가입에서 사용하는 공통 스타일 묶음
